+++ src/components/overrides/Head.astro
@@ -5,7 +5,7 @@ import { differenceInCalendarDays } from "date-fns";
 import "tippy.js/dist/tippy.css";
 
 import { getEntry } from "astro:content";
-import { getOgImage } from "~/util/props";
+import { getOgImage } from "~/util/og";
 import type { CollectionEntry } from "astro:content";
 
 // grab the current top-level folder. Remove . characters for 1.1.1.1 URL
